refactor: resources count
